Feature: add region explicitly in aws provider to avoid regressions (#15)
<!-- markdownlint-disable-next-line first-line-heading -->
## Description

This addresses the regression observed in certain configurations as
described in issue #14.

<!-- Describe your changes in detail. -->

The PR refactors the project to take the `region` from variables but
only _if and only if provided explicitly_. If not, the current behavior
is maintained. This refactor means that all modules within the project
take the provider region as a parameter which is used for all non
aliased (_i.e._ non-global) ones.

## Context

<!-- Why is this change required? What problem does it solve? -->

Resolves regressions observed in #14.

## Type of changes

<!-- What types of changes does your code introduce? Put an `x` in all
the boxes that apply. -->

- [x] Refactoring (non-breaking change)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would change existing
functionality)
- [x] Bug fix (non-breaking change which fixes an issue)

## Checklist

<!-- Go over all the following points, and put an `x` in all the boxes
that apply. -->

- [x] I am familiar with the [contributing
guidelines](../docs/CONTRIBUTING.md)
- [x] I have followed the code style of the project
- [ ] I have added tests to cover my changes
- [ ] I have updated the documentation accordingly
- [ ] This PR is a result of pair or mob programming

---

## Sensitive Information Declaration

To ensure the utmost confidentiality and protect your and others
privacy, we kindly ask you to NOT including [PII (Personal Identifiable
Information) / PID (Personal Identifiable
Data)](https://digital.nhs.uk/data-and-information/keeping-data-safe-and-benefitting-the-public)
or any other sensitive data in this PR (Pull Request) and the codebase
changes. We will remove any PR that do contain any sensitive
information. We really appreciate your cooperation in this matter.

- [x] I confirm that neither PII/PID nor sensitive data are included in
this PR and the codebase changes.

---------

Signed-off-by: Thomas Judd-Cooper <thomas.judd-cooper1@nhs.net>
Co-authored-by: Thomas Judd-Cooper <thomas.judd-cooper1@nhs.net>